+++ packages/npm/@amazeelabs/react-framework-bridge/src/__tests__/gatsby.test.tsx
@@ -1,8 +1,10 @@
-import { render, screen } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import { Field } from 'formik';
 import { GatsbyLinkProps } from 'gatsby';
 import React from 'react';
 
-import { buildLink } from '../gatsby';
+import { buildForm, buildLink } from '../gatsby';
 
 const gatsbyNav = jest.fn();
 
@@ -22,6 +24,7 @@ jest.mock(
     navigate: (to: string) => gatsbyNav(to),
   }),
 );
+beforeEach(jest.resetAllMocks);
 
 describe('buildLink', () => {
   it('can build a link from segments and query parameters', () => {
@@ -31,6 +34,7 @@ describe('buildLink', () => {
         a: 'b',
       },
     });
+    expect(Link.href).toEqual('/foo/bar?a=b');
     render(<Link>Test</Link>);
     expect(screen.getByRole('link').getAttribute('data-gatsby')).toBeTruthy();
     expect(screen.getByRole('link').getAttribute('href')).toEqual(
@@ -38,14 +42,61 @@ describe('buildLink', () => {
     );
   });
 
+  it('can build a link from only query parameters', () => {
+    const Link = buildLink({
+      query: {
+        a: 'b',
+      },
+    });
+    expect(Link.href).toEqual('?a=b');
+    render(<Link>Test</Link>);
+    expect(screen.getByRole('link').getAttribute('data-gatsby')).toBeTruthy();
+    expect(screen.getByRole('link').getAttribute('href')).toEqual('?a=b');
+  });
+
+  it('allows the consumer to set query parameters', () => {
+    const Link = buildLink({
+      href: '/foo',
+    });
+    expect(Link.href).toEqual('/foo');
+    render(<Link query={{ a: 'b' }}>Test</Link>);
+    expect(screen.getByRole('link').getAttribute('href')).toEqual('/foo?a=b');
+  });
+
+  it('allows the consumer to set a query fragment', () => {
+    const Link = buildLink({
+      href: '/foo',
+    });
+    expect(Link.href).toEqual('/foo');
+    render(<Link fragment="bar">Test</Link>);
+    expect(screen.getByRole('link').getAttribute('href')).toEqual('/foo#bar');
+  });
+
+  it('allows the consumer to set query parameters and fragments', () => {
+    const Link = buildLink({
+      href: '/foo',
+    });
+    expect(Link.href).toEqual('/foo');
+    render(
+      <Link query={{ a: 'b' }} fragment="bar">
+        Test
+      </Link>,
+    );
+    expect(screen.getByRole('link').getAttribute('href')).toEqual(
+      '/foo?a=b#bar',
+    );
+  });
+
   it('renders a Gatsby link for an internal path', () => {
     const Link = buildLink({ href: '/test' });
+    expect(Link.href).toEqual('/test');
     render(<Link>Test</Link>);
     expect(screen.getByRole('link').getAttribute('data-gatsby')).toBeTruthy();
   });
 
   it('renders normal link for an external path', () => {
     const Link = buildLink({ href: 'http://www.amazeelabs.com' });
+    expect(Link.href).toEqual('http://www.amazeelabs.com');
     render(<Link>Test</Link>);
     expect(screen.getByRole('link').getAttribute('data-gatsby')).toBeFalsy();
   });
@@ -55,6 +106,7 @@ describe('buildLink', () => {
       href: 'http://www.amazeelabs.com',
       target: '_blank',
     });
+    expect(Link.href).toEqual('http://www.amazeelabs.com');
     render(<Link>Test</Link>);
     expect(screen.getByRole('link').getAttribute('data-gatsby')).toBeFalsy();
   });
@@ -63,14 +115,46 @@ describe('buildLink', () => {
     const Link = buildLink({
       href: 'mailto:development@amazeelabs.com',
     });
+    expect(Link.href).toEqual('mailto:development@amazeelabs.com');
     render(<Link>Test</Link>);
     expect(screen.getByRole('link').getAttribute('data-gatsby')).toBeFalsy();
   });
 
   it('exposes Gatsby navigate', () => {
     const Link = buildLink({ href: '#test' });
+    expect(Link.href).toEqual('#test');
     Link.navigate();
     expect(gatsbyNav).toHaveBeenCalledTimes(1);
     expect(gatsbyNav).toHaveBeenCalledWith('#test');
+  });
+
+  it('exposes Gatsby navigate that allows to override query and fragments', () => {
+    const Link = buildLink({ href: '/foo', query: { a: 'b' } });
+    expect(Link.href).toEqual('/foo?a=b');
+    Link.navigate({ query: { a: 'c' }, fragment: 'bar' });
+    expect(gatsbyNav).toHaveBeenCalledTimes(1);
+    expect(gatsbyNav).toHaveBeenCalledWith('/foo?a=c#bar');
+  });
+});
+
+describe('buildForm', () => {
+  it('hands form values to the submit callback', async () => {
+    const callback = jest.fn();
+    const Form = buildForm({
+      initialValues: { foo: '' },
+      onSubmit: (values) => callback(values),
+    });
+    render(
+      <Form>
+        <Field type="text" name="foo" />
+        <button type="submit" />
+      </Form>,
+    );
+    userEvent.type(screen.getByRole('textbox'), 'bar');
+    userEvent.click(screen.getByRole('button'));
+    await waitFor(() => {
+      expect(callback).toHaveBeenCalledTimes(1);
+      expect(callback).toHaveBeenCalledWith({ foo: 'bar' });
+    });
   });
 });