+++ packages/npm/@amazeelabs/silverback-playwright/src/drupal-only/global-setup.ts
@@ -1,11 +1,13 @@
 import { $, cd, fs } from 'zx';
 
 import { drupal } from '../constants';
-import { port } from '../utils';
+import { log, port } from '../utils';
 
 $.verbose = !!process.env.SP_VERBOSE;
 
 export default async function globalSetup() {
+  log('drupal-only globalSetup start');
+
   cd(drupal.path);
 
   if (fs.existsSync(`${drupal.path}/.silverback-snapshots/test`)) {
@@ -18,4 +20,6 @@ export default async function globalSetup() {
   await port.killIfUsed(drupal.port);
   $`TEST_SESSION_ENABLED=true yarn start`;
   await port.waitUntilUsed(drupal.port);
+
+  log('drupal-only globalSetup end');
 }