+++ packages/npm/@amazeelabs/silverback-playwright/src/test-package.ts
@@ -10,13 +10,15 @@ if (process.argv.includes('help') || process.argv.includes('--help')) {
   console.log(`Usage: yarn sp-test [flags]
   Flags:
     -v, --verbose    Spam a lot
-    -h, --headed     Run tests in headed browser (default: headless)`);
+    -h, --headed     Run tests in headed browser (default: headless)
+    -t, --trace      Record traces`);
   process.exit();
 }
 
 const headed = process.argv.includes('--headed') || process.argv.includes('-h');
 const verbose =
   process.argv.includes('--verbose') || process.argv.includes('-v');
+const trace = process.argv.includes('--trace') || process.argv.includes('-t');
 
 $.verbose = verbose;
 // We escape args ourselves.
@@ -27,6 +29,7 @@ const runTests = async (type: TestType) => {
     `SP_TEST_DIR='${process.cwd()}/playwright-tests'`,
     `SP_TEST_TYPE=${type}`,
     `SP_VERBOSE=${verbose ? 'true' : "''"}`,
+    `SP_TRACE=${trace ? 'true' : "''"}`,
   ].join(' ');
 
   console.log(`⏩ Running ${type} tests...`);
@@ -35,7 +38,7 @@ const runTests = async (type: TestType) => {
     $.verbose = true;
     const run = $`${envVars} yarn playwright test --grep '${
       tags[type]
-    }' --config '${__dirname}/${type}/playwright.config.ts' ${
+    }' --config '${__dirname}/${type}/playwright.config.ts' --workers 1 --max-failures 1 ${
       headed ? '--headed --timeout 600000' : ''
     }`;
     $.verbose = verbose;