+++ apps/silverback-drupal/web/modules/custom/gutenberg_custom_blocks/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.7](https://github.com/AmazeeLabs/silverback-mono/compare/@-silverback-drupal/gutenberg_custom_blocks@1.0.6...@-silverback-drupal/gutenberg_custom_blocks@1.0.7) (2021-10-04)
+
+**Note:** Version bump only for package @-silverback-drupal/gutenberg_custom_blocks
+
+
+
+
+
+## [1.0.6](https://github.com/AmazeeLabs/silverback-mono/compare/@-silverback-drupal/gutenberg_custom_blocks@1.0.5...@-silverback-drupal/gutenberg_custom_blocks@1.0.6) (2021-09-28)
+
+**Note:** Version bump only for package @-silverback-drupal/gutenberg_custom_blocks
+
+
+
+
+
+## [1.0.5](https://github.com/AmazeeLabs/silverback-mono/compare/@-silverback-drupal/gutenberg_custom_blocks@1.0.4...@-silverback-drupal/gutenberg_custom_blocks@1.0.5) (2021-09-28)
+
+**Note:** Version bump only for package @-silverback-drupal/gutenberg_custom_blocks
+
+
+
+
+
+## [1.0.4](https://github.com/AmazeeLabs/silverback-mono/compare/@-silverback-drupal/gutenberg_custom_blocks@1.0.3...@-silverback-drupal/gutenberg_custom_blocks@1.0.4) (2021-09-15)
+
+**Note:** Version bump only for package @-silverback-drupal/gutenberg_custom_blocks
+
+
+
+
+
 ## [1.0.3](https://github.com/AmazeeLabs/silverback-mono/compare/@-silverback-drupal/gutenberg_custom_blocks@1.0.2...@-silverback-drupal/gutenberg_custom_blocks@1.0.3) (2021-09-06)
 
 **Note:** Version bump only for package @-silverback-drupal/gutenberg_custom_blocks