+++ packages/npm/@amazeelabs/gatsby-source-silverback-cypress/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.8](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback-cypress@1.1.7...@amazeelabs/gatsby-source-silverback-cypress@1.1.8) (2021-10-04)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-source-silverback-cypress
+
+
+
+
+
+## [1.1.7](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback-cypress@1.1.6...@amazeelabs/gatsby-source-silverback-cypress@1.1.7) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-source-silverback-cypress
+
+
+
+
+
 ## [1.1.6](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback-cypress@1.1.5...@amazeelabs/gatsby-source-silverback-cypress@1.1.6) (2021-09-06)
 
 **Note:** Version bump only for package @amazeelabs/gatsby-source-silverback-cypress