+++ packages/npm/@amazeelabs/gatsby-source-silverback/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.9](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback@1.4.8...@amazeelabs/gatsby-source-silverback@1.4.9) (2021-10-04)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-source-silverback
+
+
+
+
+
+## [1.4.8](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback@1.4.7...@amazeelabs/gatsby-source-silverback@1.4.8) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-source-silverback
+
+
+
+
+
+## [1.4.7](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback@1.4.6...@amazeelabs/gatsby-source-silverback@1.4.7) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-source-silverback
+
+
+
+
+
+## [1.4.6](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback@1.4.5...@amazeelabs/gatsby-source-silverback@1.4.6) (2021-09-20)
+
+
+### Bug Fixes
+
+* move silverback-playwright to dev dependencies ([23e90a9](https://github.com/AmazeeLabs/silverback-mono/commit/23e90a9340fe243ba4cb469d3b5985a688374f20))
+
+
+
+
+
+## [1.4.5](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback@1.4.4...@amazeelabs/gatsby-source-silverback@1.4.5) (2021-09-15)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-source-silverback
+
+
+
+
+
+## [1.4.4](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback@1.4.3...@amazeelabs/gatsby-source-silverback@1.4.4) (2021-09-15)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-source-silverback
+
+
+
+
+
 ## [1.4.3](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-source-silverback@1.4.2...@amazeelabs/gatsby-source-silverback@1.4.3) (2021-09-06)
 
 **Note:** Version bump only for package @amazeelabs/gatsby-source-silverback