+++ packages/npm/@amazeelabs/jest-preset/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.7](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/jest-preset@1.3.6...@amazeelabs/jest-preset@1.3.7) (2021-10-04)
+
+
+### Bug Fixes
+
+* **deps:** update dependency jest-watch-typeahead to v1 ([ee0e842](https://github.com/AmazeeLabs/silverback-mono/commit/ee0e842624907615ec7a33629346abad6448fe88))
+
+
+
+
+
+## [1.3.6](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/jest-preset@1.3.5...@amazeelabs/jest-preset@1.3.6) (2021-09-28)
+
+
+### Bug Fixes
+
+* **deps:** update dependency jest-runner-eslint to ^0.11.0 ([967ebfc](https://github.com/AmazeeLabs/silverback-mono/commit/967ebfcce593b5cc0dc7e0cf682f9e491903974f))
+
+
+
+
+
 ## [1.3.5](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/jest-preset@1.3.4...@amazeelabs/jest-preset@1.3.5) (2021-09-06)
 
 **Note:** Version bump only for package @amazeelabs/jest-preset