+++ packages/npm/@amazeelabs/ui/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.14](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/ui@0.2.13...@amazeelabs/ui@0.2.14) (2021-09-15)
+
+**Note:** Version bump only for package @amazeelabs/ui
+
+
+
+
+
 ## [0.2.13](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/ui@0.2.12...@amazeelabs/ui@0.2.13) (2021-07-29)
 
 **Note:** Version bump only for package @amazeelabs/ui