+++ packages/npm/drupal-test-session-cypress/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.11](https://github.com/AmazeeLabs/silverback-mono/compare/drupal-test-session-cypress@1.1.10...drupal-test-session-cypress@1.1.11) (2021-10-04)
+
+**Note:** Version bump only for package drupal-test-session-cypress
+
+
+
+
+
+## [1.1.10](https://github.com/AmazeeLabs/silverback-mono/compare/drupal-test-session-cypress@1.1.9...drupal-test-session-cypress@1.1.10) (2021-09-28)
+
+**Note:** Version bump only for package drupal-test-session-cypress
+
+
+
+
+
 ## [1.1.9](https://github.com/AmazeeLabs/silverback-mono/compare/drupal-test-session-cypress@1.1.8...drupal-test-session-cypress@1.1.9) (2021-09-06)
 
 **Note:** Version bump only for package drupal-test-session-cypress